+++ backend/Musify.Infrastructure/MassTransit/Logs/UpdateTrackAudioLog.cs
@@ -3,5 +3,5 @@
     public record UpdateTrackAudioLog(
         Guid TrackId,
         string? PreviousAudioFolderName,
-        int PreviousDurationSeconds);
+        double PreviousDurationSeconds);
 }
